ai-service/main.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Optional
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(title="Orbix AI Service")

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Models
class ContentGenerationRequest(BaseModel):
    prompt: str
    type: str  # blog, email, social, etc.
    tone: Optional[str] = "professional"
    length: Optional[str] = "medium"

class GrammarCheckRequest(BaseModel):
    text: str

class RewriteRequest(BaseModel):
    text: str
    tone: str  # professional, casual, formal, friendly

class SummarizeRequest(BaseModel):
    text: str
    length: Optional[str] = "short"

class ResumeOptimizeRequest(BaseModel):
    resume_text: str
    job_description: str

class ContractGenerateRequest(BaseModel):
    contract_type: str
    parties: List[dict]
    terms: dict

# Routes
@app.get("/health")
async def health():
    return {"status": "ok", "service": "Orbix AI Service"}

@app.post("/api/ai/generate-content")
async def generate_content(request: ContentGenerationRequest):
    # TODO: Implement OpenAI integration
    return {
        "content": f"Generated {request.type} content based on: {request.prompt}",
        "tokens_used": 0
    }

@app.post("/api/ai/check-grammar")
async def check_grammar(request: GrammarCheckRequest):
    # TODO: Implement grammar checking
    return {
        "corrected_text": request.text,
        "errors": [],
        "suggestions": []
    }

@app.post("/api/ai/rewrite")
async def rewrite_text(request: RewriteRequest):
    # TODO: Implement text rewriting
    return {
        "rewritten_text": request.text,
        "tone": request.tone
    }

@app.post("/api/ai/summarize")
async def summarize_text(request: SummarizeRequest):
    # TODO: Implement summarization
    return {
        "summary": f"Summary of the text (length: {request.length})"
    }

@app.post("/api/ai/optimize-resume")
async def optimize_resume(request: ResumeOptimizeRequest):
    # TODO: Implement resume optimization
    return {
        "optimized_resume": request.resume_text,
        "ats_score": 85,
        "suggestions": [
            "Add more action verbs",
            "Include relevant keywords from job description",
            "Quantify achievements with metrics"
        ],
        "missing_keywords": ["Python", "AWS", "Docker"]
    }

@app.post("/api/ai/generate-contract")
async def generate_contract(request: ContractGenerateRequest):
    # TODO: Implement contract generation
    return {
        "contract_content": f"Generated {request.contract_type} contract",
        "sections": []
    }

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
