ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.10.0
langchain==0.1.0
pydantic==2.5.3
spacy==3.7.2
PyMuPDF==1.23.8
python-docx==1.1.0
